[veSync] Move common class
[veSync] Move common class
