+++ bundles/org.openhab.binding.vesync/src/main/java/org/openhab/binding/vesync/internal/handlers/VeSyncDeviceAirHumidifierHandler.java
@@ -29,9 +29,9 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
 import org.openhab.binding.vesync.internal.VeSyncBridgeConfiguration;
 import org.openhab.binding.vesync.internal.VeSyncConstants;
+import org.openhab.binding.vesync.internal.dto.requests.v2.EmptyPayload;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev1.SetNightLightBrightness;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev1.SetTargetHumidity;
-import org.openhab.binding.vesync.internal.dto.requests.v2.rev2.EmptyPayload;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev2.Enabled;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev2.SetLevel;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev2.SetMode;
