+++ bundles/org.openhab.binding.vesync/src/main/java/org/openhab/binding/vesync/internal/handlers/VeSyncDeviceAirPurifierHandler.java
@@ -34,8 +34,8 @@ import org.openhab.binding.vesync.internal.dto.requests.v1.GetDeviceDetails;
 import org.openhab.binding.vesync.internal.dto.requests.v1.SetLevel;
 import org.openhab.binding.vesync.internal.dto.requests.v1.SetMode;
 import org.openhab.binding.vesync.internal.dto.requests.v1.SetStatus;
+import org.openhab.binding.vesync.internal.dto.requests.v2.EmptyPayload;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev1.SetNightLight;
-import org.openhab.binding.vesync.internal.dto.requests.v2.rev2.EmptyPayload;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev2.SetChildLock;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev2.SetLightDetection;
 import org.openhab.binding.vesync.internal.dto.requests.v2.rev2.SetManualSpeedLevel;
